Update rebar.config.script
Suprisingly, removing this line fixes a bug with autocomplete in emacs, specifically with elixir + lsp + company mode.

This happens on M1 Mac and Android. This line is logged everytime eradius starts, which is annoying at best, but ends up prepending that line to every file saved from emacs (unless autocomplete is disabled).